--- src/helpers/index.js
@@ -1,7 +1,0 @@
-import * as ObjectHelper from './ObjectHelper/ObjectHelper';
-import * as TimeHelper from './TimeHelper/TimeHelper';
-
-export {
-    ObjectHelper,
-    TimeHelper
-};
